ics/requirements.txt
Click==7.0
ics==0.4
python-dateutil==2.8.0
